+++ mpire/__init__.py
@@ -1,3 +1,3 @@
 from .mpire import cpu_count, Worker, WorkerPool
 
-__version__ = '0.5.1'
+__version__ = '0.6.0'