+++ mpire/mpire.py
@@ -1,9 +1,13 @@
 import itertools
+import os
 import queue
+import signal
 import subprocess
+import time
 import tqdm
+import traceback
 import warnings
-from multiprocessing import cpu_count, Event, Process, Queue
+from multiprocessing import cpu_count, Event, JoinableQueue, Lock, Process, Value
 from threading import Thread
 
 
@@ -20,13 +24,19 @@ except NameError:
     from tqdm import tqdm as tqdm_pb
 
 
+class StopWorker(Exception):
+    """ Exception used to kill workers from the main process """
+    pass
+
+
 class Worker(Process):
     """
     A multiprocessing helper class which continuously asks the queue for new jobs, until a poison pill is inserted
     """
 
-    def __init__(self, worker_id, tasks_queue, results_queue, restart_queue, task_completed_queue, func_pointer,
-                 keep_order_event, shared_objects=None, worker_lifespan=None, pass_worker_id=False):
+    def __init__(self, worker_id, tasks_queue, results_queue, restart_queue, task_completed_queue, exception_queue,
+                 has_started_counter, func_pointer, keep_order_event, shared_objects=None, worker_lifespan=None,
+                 pass_worker_id=False):
         """
         :param worker_id: Worker id
         :param tasks_queue: Queue object for retrieving new task arguments
@@ -34,6 +44,9 @@ class Worker(Process):
         :param restart_queue: Queue object for notifying the ``WorkerPool`` to restart this worker
         :param task_completed_queue: Queue object for notifying the main process we're done with a single task. ``None``
             when notifying is not necessary.
+        :param exception_queue: Queue object for sending Exception objects to whenever an Exception was raised inside a
+            user function
+        :param has_started_counter: multiprocessing.Value object that tracks how many workers have started
         :param func_pointer: Function pointer to call each time new task arguments become available
         :param keep_order_event: Event object which signals if the task arguments contain an order index which should
             be preserved and not fed to the function pointer (e.g., used in ``map``)
@@ -45,63 +58,150 @@ class Worker(Process):
         :param pass_worker_id: Boolean. Whether or not to pass the worker ID to the function
         """
         super().__init__()
+
+        # Parameters
         self.worker_id = worker_id
         self.tasks_queue = tasks_queue
         self.results_queue = results_queue
         self.restart_queue = restart_queue
         self.task_completed_queue = task_completed_queue
+        self.exception_queue = exception_queue
+        self.has_started_counter = has_started_counter
         self.func_pointer = func_pointer
         self.keep_order_event = keep_order_event
         self.shared_objects = shared_objects
         self.worker_lifespan = worker_lifespan
         self.pass_worker_id = pass_worker_id
 
+        # Exception handling variables
+        self.running = False
+        self.lock = Lock()
+        self.stop_event = Event()
+
+        # Register handler for graceful shutdown
+        signal.signal(signal.SIGUSR1, self._exit_gracefully)
+
+    def _exit_gracefully(self, *_):
+        """
+        This function is called when the main process sends a kill signal to this process. This can only mean another
+        child process encountered an exception which means we should exit.
+
+        We do this by setting the stop event and, if the process is in the middle of running the user defined function,
+        by raising a StopWorker exception (which is then caught by the ``run()`` function) so we can quit gracefully.
+        """
+        # A rather complex locking mechanism is used here so we can make sure we only raise an exception when we should.
+        # We want to make sure we only raise an exception when the user function is called. If it is not running, the
+        # stop event will make sure `run` will return. However, this may be a bit delayed because we could be in the
+        # middle of sending items through the queue. If it is running the exception is caught and `run` will return.
+        # If it is running and the user function throws another exception first, it depends on who obtains the lock
+        # first. If `run` obtains it it will set `running` to False, meaning we won't raise and `run` will return. If
+        # this function obtains it first it will throw, which again is caught by the `run` function, which will return.
+        # It should be bullet proof :)
+        with self.lock:
+            self.stop_event.set()
+            if self.running:
+                raise StopWorker
+
     def run(self):
         """
         Continuously asks the tasks queue for new task arguments. When not receiving a poisonous pill or when the max
         life span is not yet reached it will execute the new task and put the results in the results queue.
         """
+        # Determine what function to call. If we have to keep in mind the order (for map) we use the helper function
+        # which deals with the provided idx variable.
+        if self.keep_order_event.is_set():
+            def func(_args, _additional_args):
+                return self._helper_func(*itertools.chain(_args, [_additional_args]))
+
+        # In the case of not keeping order in mind (for imap) we can simply call the user function directly
+        else:
+            def func(_args, _additional_args):
+                return self.func_pointer(*itertools.chain(_additional_args, _args))
+
+        # Obtain additional args to pass to the function
+        additional_args = []
+        if self.pass_worker_id:
+            additional_args.append(self.worker_id)
+        if self.shared_objects is not None:
+            additional_args.append(self.shared_objects)
+
         n_chunks_executed = 0
         while self.worker_lifespan is None or n_chunks_executed < self.worker_lifespan:
-            # Obtain new job
-            next_chunked_args = self.tasks_queue.get()
-            if next_chunked_args is None:
-                # Poison pill means we should exit
-                break
-
-            # Function to call
-            if self.keep_order_event.is_set():
-                def func(_args, _additional_args):
-                    return self._helper_func(*itertools.chain(_args, [_additional_args]))
+            # Obtain new chunk of jobs and occasionally poll the stop event
+            while not self.stop_event.is_set():
+                try:
+                    next_chunked_args = self.tasks_queue.get(block=True, timeout=0.1)
+                    break
+                except queue.Empty:
+                    pass
             else:
-                def func(_args, _additional_args):
-                    return self.func_pointer(*itertools.chain(_additional_args, _args))
+                # Stop event is set, we should shut down
+                return
 
-            # Obtain additional args to pass to the function
-            additional_args = []
-            if self.pass_worker_id:
-                additional_args.append(self.worker_id)
-            if self.shared_objects is not None:
-                additional_args.append(self.shared_objects)
+            # If we obtained a poison pill, we stop
+            if next_chunked_args is '\0':
+                self.tasks_queue.task_done()
+                return
 
-            # Execute jobs
+            # Execute jobs in this chunk
             results = []
             for args in next_chunked_args:
-                results.append(func(args, additional_args))
+                try:
+                    # Set running to True, so when we need to shut down the signal handler knows it can raise
+                    with self.lock:
+                        self.running = True
 
-                # Notify that we've completed a task
+                        # Before every task, we also check the stop event. This is needed when the kill signal is given
+                        # in this loop, right after we have set running to False. In that case the signal handler will
+                        # not raise, but we would be still running tasks.
+                        if self.stop_event.is_set():
+                            # print(self.worker_id, "stop event is set right before processing, shutting down")
+                            self.tasks_queue.task_done()
+                            return
+
+                    # This can fail if an exception occurs in the user provided function
+                    try:
+                        results.append(func(args, additional_args))
+                    except StopWorker:
+                        raise
+                    except Exception as err:
+                        # Let the signal handler know it shouldn't raise any exception anymore, we got this.
+                        with self.lock:
+                            self.running = False
+
+                        # Create traceback string and add exception to queue
+                        traceback_str = "\n\nException occurred in Worker-%d with the following arguments:\n%s\n%s" % \
+                                        (self.worker_id,
+                                         "\n".join("Arg %d: %s" % (arg_nr, str(arg))
+                                                   for arg_nr, arg in
+                                                   enumerate(args[1] if self.keep_order_event.is_set() else args)),
+                                         traceback.format_exc())
+                        self.exception_queue.put((type(err), traceback_str), block=False, timeout=2)
+                        self.tasks_queue.task_done()
+                        return
+
+                    # Set running to False, so when we need to shut down the signal handler knows it should only set the
+                    # stop event
+                    with self.lock:
+                        self.running = False
+
+                except StopWorker:
+                    # The main process tells us to stop working, shutting down
+                    self.tasks_queue.task_done()
+                    return
+
+                # Notify that we've completed a task (only when using a progress bar)
                 if self.task_completed_queue is not None:
                     self.task_completed_queue.put(1)
 
             # Send results back to main process
             self.results_queue.put(results)
-
-            # Increment counter
+            self.tasks_queue.task_done()
             n_chunks_executed += 1
 
         # Notify WorkerPool to start a new worker if max lifespan is reached
         if self.worker_lifespan is not None and n_chunks_executed == self.worker_lifespan:
-            self.restart_queue.put(self.worker_id)
+            self.restart_queue.put(self.worker_id, block=True)
 
     def _helper_func(self, idx, args, additional_args):
         """
@@ -136,17 +236,55 @@ class WorkerPool:
         self.daemon = daemon
         self.cpu_ids = self._check_cpu_ids(cpu_ids)
 
-        # Containers for later use
-        self.tasks_queue = None
-        self.results_queue = None
-        self.restart_queue = None
-        self.task_completed_queue = None
-        self.keep_order_event = Event()
-        self.workers = []
+        # Queue to pass on tasks to child processes
+        self._tasks_queue = None
+
+        # Queue where the child processes can pass on results
+        self._results_queue = None
+
+        # Queue where the child processes can request a restart
+        self._restart_queue = None
+
+        # Queue related to the progress bar. Child processes can pass on a random value whenever they are finished with
+        # a job
+        self._task_completed_queue = None
+
+        # Queue where the child processes can pass on an encountered exceptoin
+        self._exception_queue = None
+
+        # Synchronized value counting how many child processes actually started (when calling .start() on a Process does
+        # not mean that process is actually spawned yet)
+        self._has_started_counter = None
+
+        # Synchronized event to inform the child processes to keep order in mind (for the map functions)
+        self._keep_order_event = Event()
+
+        # Container of the child processes
+        self._workers = []
+
+        # User provided shared objects to pass on to the user provided function
         self.shared_objects = None
+
+        # User provided function to call
         self.func_pointer = None
+
+        # Number of (chunks of) jobs a child process can process before requesting a restart
         self.worker_lifespan = None
+
+        # Whether to pass on the worker ID to the user provided function
         self.pass_worker_id = False
+
+        # Whether or not an exception was caught by one of the child processes
+        self.exception_caught = False
+
+        # Thread that handles the progress bar updates
+        self._progress_bar_thread = None
+
+        # Thread that handles incoming exceptions from the child processes
+        self._exception_thread = None
+
+        # Synchronized event indicating whether or not the main process is ready for destruction in case of an exception
+        self._ready_for_destruction = Event()
 
     def _check_cpu_ids(self, cpu_ids):
         """
@@ -243,63 +381,65 @@ class WorkerPool:
             ``None``, workers will stay alive the entire time. Use this when workers use up too much memory over the
             course of time.
         """
-        # If there are workers, join them first
-        self.stop_and_join()
-
-        # If worker lifespan is not None or not a positive integer, raise
-        if worker_lifespan is not None and not (isinstance(worker_lifespan, int) and worker_lifespan > 0):
-            raise ValueError('worker_lifespan should be either None or a positive integer (> 0)')
-
         # Save params for later reference (for example, when restarting workers)
         self.func_pointer = func_pointer
         self.worker_lifespan = worker_lifespan
 
         # Start new workers
-        self.tasks_queue = Queue()
-        self.results_queue = Queue()
-        self.restart_queue = Queue()
+        self._tasks_queue = JoinableQueue()
+        self._results_queue = JoinableQueue()
+        self._restart_queue = JoinableQueue()
+        self._exception_queue = JoinableQueue()
+        self._has_started_counter = Value('i', 0, lock=True)
         for worker_id in range(self.n_jobs):
-            w = Worker(worker_id, self.tasks_queue, self.results_queue, self.restart_queue, self.task_completed_queue,
-                       self.func_pointer, self.keep_order_event, self.shared_objects, self.worker_lifespan,
-                       self.pass_worker_id)
-            w.daemon = self.daemon
-            w.start()
-            if self.cpu_ids:
-                subprocess.call('taskset -p -c %s %d' % (self.cpu_ids[worker_id], w.pid), stdout=subprocess.DEVNULL,
-                                shell=True)
-            self.workers.append(w)
+            self._workers.append(self._start_worker(worker_id))
 
     def _restart_workers(self):
         """
         Restarts workers that need to be restarted.
         """
-        while True:
+        while not self.exception_caught:
             try:
-                # Obtain worker id that needs to be restarted
-                worker_id = self.restart_queue.get(block=False)
-                self.workers[worker_id].join()
+                # Obtain worker ID that needs to be restarted and join the process
+                worker_id = self._restart_queue.get(block=False)
+                self._workers[worker_id].join()
+                self._has_started_counter.value -= 1
 
                 # Start worker
-                w = Worker(worker_id, self.tasks_queue, self.results_queue, self.restart_queue,
-                           self.task_completed_queue, self.func_pointer, self.keep_order_event, self.shared_objects,
-                           self.worker_lifespan, self.pass_worker_id)
-                w.daemon = self.daemon
-                w.start()
-                if self.cpu_ids:
-                    subprocess.call('taskset -p -c %s %d' % (self.cpu_ids[worker_id], w.pid), stdout=subprocess.DEVNULL,
-                                    shell=True)
-                self.workers[worker_id] = w
+                self._workers[worker_id] = self._start_worker(worker_id)
+                self._restart_queue.task_done()
             except queue.Empty:
                 # There are no workers to be restarted, we're done here
                 break
+
+    def _start_worker(self, worker_id):
+        """
+        Creates and starts a single worker
+
+        :param worker_id: ID of the worker
+        :return: Worker instance
+        """
+        # Create worker
+        w = Worker(worker_id, self._tasks_queue, self._results_queue, self._restart_queue, self._task_completed_queue,
+                   self._exception_queue, self._has_started_counter, self.func_pointer, self._keep_order_event,
+                   self.shared_objects, self.worker_lifespan, self.pass_worker_id)
+        w.daemon = self.daemon
+        w.start()
+
+        # Pin CPU if desired
+        if self.cpu_ids:
+            subprocess.call('taskset -p -c %s %d' % (self.cpu_ids[worker_id], w.pid), stdout=subprocess.DEVNULL,
+                            shell=True)
+
+        return w
 
     def add_task(self, args):
         """
         Add a task to the queue so a worker can process it.
 
-        :param args: A tuple of arguments to pass to a worker, which passes it to the function pointer
+        :param args: A tuple of arguments to pass to a worker, which acts upon it
         """
-        self.tasks_queue.put(args)
+        self._tasks_queue.put(args)
 
     def get_result(self):
         """
@@ -307,14 +447,16 @@ class WorkerPool:
 
         :return: Various. The next result from the queue, which is the result of calling the function pointer.
         """
-        return self.results_queue.get(block=True)
+        results = self._results_queue.get(block=True)
+        self._results_queue.task_done()
+        return results
 
     def insert_poison_pill(self):
         """
         Tell the workers their job is done by killing them brutally.
         """
-        for _ in range(len(self.workers)):
-            self.tasks_queue.put(None)
+        for _ in self._workers:
+            self.add_task('\0')
 
     stop_workers = insert_poison_pill
 
@@ -326,7 +468,7 @@ class WorkerPool:
         For more information, see the warnings at:
         https://docs.python.org/3.4/library/multiprocessing.html#pipes-and-queues.
         """
-        for w in self.workers:
+        for w in self._workers:
             w.join()
 
     def stop_and_join(self):
@@ -337,22 +479,146 @@ class WorkerPool:
         For more information, see the warnings at:
         https://docs.python.org/3.4/library/multiprocessing.html#pipes-and-queues.
         """
-        if self.workers:
+        if self._workers:
+            # Insert poison pill
             self.insert_poison_pill()
-            self.join()
-            self.workers = []
-            self.tasks_queue = None
-            self.results_queue = None
 
-    def terminate(self):
+            # It can occur that some processes requested a restart while all tasks are already complete. This means that
+            # processes won't get restarted anymore, although a poison pill is just added (at the time they we're still
+            # barely alive). This results in a tasks queue never being empty which can lead to BrokenPipeErrors or
+            # deadlocks. We can also not simply drain the tasks queue because workers could still be busy. To remedy
+            # this problem we start a thread which waits for the tasks queue (including poison pills) to be empty. If
+            # this thread isn't done that means that some tasks/poison pills are still there, meaning that there are
+            # child processes which could/should be restarted, but aren't.
+            t = Thread(target=self._handle_tasks_done)
+            t.start()
+            while True:
+                # If not done, try restarting workings
+                t.join(timeout=0.1)
+                if not t.is_alive():
+                    break
+                self._restart_workers()
+
+            # Join workers
+            self.join()
+
+            # Join queues
+            self._tasks_queue.join()
+            self._results_queue.join()
+            self._restart_queue.join()
+            if self._task_completed_queue is not None:
+                self._task_completed_queue.join()
+
+            # Reset variables
+            self._workers = []
+            self._tasks_queue = None
+            self._results_queue = None
+            self._restart_queue = None
+            self._task_completed_queue = None
+            self._exception_queue = None
+            self._has_started_counter = None
+
+    def _handle_tasks_done(self):
         """
-        Does not wait until all workers are finished, but terminates them with a SIGTERM.
+        Blocks until the tasks queue is joinable
         """
-        for w in self.workers:
-            w.terminate()
-            self.workers = []
-            self.tasks_queue = None
-            self.results_queue = None
+        self._tasks_queue.join()
+
+    def terminate(self, keep_exception_queue=False):
+        """
+        Does not wait until all workers are finished, but terminates them.
+
+        :param keep_exception_queue: Whether to keep the exception queue or not
+        """
+        # Send interrupt signal so the processes can die gracefully. Sometimes this will throw an exception, e.g. when
+        # processes are in the middle of restarting
+        for w in self._workers:
+            try:
+                os.kill(w.pid, signal.SIGUSR1)
+            except ProcessLookupError:
+                pass
+
+        # When the kill signal is send to all child processes we check one process at a time if we can join them
+        for w in self._workers:
+            # We wait until workers are done terminating. However, we don't have all the patience in the world and
+            # sometimes workers can get stuck when starting up when an exception is handled. So when the patience runs
+            # out we terminate them.
+            try_count = 10
+            while w.is_alive() and try_count > 0:
+                time.sleep(0.1)
+                try:
+                    os.kill(w.pid, signal.SIGUSR1)
+                except ProcessLookupError:
+                    pass
+                try_count -= 1
+
+            # If a graceful kill is not possible, terminate the process.
+            if try_count == 0:
+                w.terminate()
+
+            # Join worker
+            w.join()
+
+        # Drain and join the queues
+        self._drain_and_join_queue(self._tasks_queue)
+        self._drain_and_join_queue(self._results_queue)
+        self._drain_and_join_queue(self._restart_queue)
+
+        # Make sure the progress bar handle thread is dead
+        if self._progress_bar_thread is not None and self._progress_bar_thread.is_alive():
+            self._drain_and_join_queue(self._task_completed_queue, join=False)
+            self._task_completed_queue.put(None)
+            self._task_completed_queue.join()
+            try:
+                self._progress_bar_thread.join()
+            except AttributeError:
+                pass
+        self._progress_bar_thread = None
+
+        # Make sure the exception bar handle thread is dead
+        if not keep_exception_queue:
+            if self._exception_thread is not None and self._exception_thread.is_alive():
+                self._drain_and_join_queue(self._exception_queue, join=False)
+                self._exception_queue.put((None, None))
+                self._exception_queue.join()
+                self._exception_thread.join()
+            self._exception_thread = None
+            self._exception_queue = None
+
+        # Reset variables
+        self._workers = []
+        self._tasks_queue = None
+        self._results_queue = None
+        self._restart_queue = None
+        self._task_completed_queue = None
+        self._has_started_counter = None
+
+    @staticmethod
+    def _drain_and_join_queue(q, join=True):
+        """
+        Drains a queue completely, such that it is joinable
+
+        :param join: Whether to join the queue or not
+        """
+        # Extract all items left in the queue
+        try:
+            while True:
+                q.get(block=False)
+                q.task_done()
+        except (queue.Empty, EOFError):
+            pass
+
+        # Call task done up to the point where we get a ValueError. We need to do this when child processes already
+        # started processing on some tasks and got terminated half-way.
+        try:
+            while True:
+                q.task_done()
+        except ValueError:
+            pass
+
+        # Join
+        if join:
+            q.join()
 
     def __enter__(self):
         """
@@ -362,16 +628,13 @@ class WorkerPool:
 
     def __exit__(self, *_):
         """
-        Enable the use of the ``with`` statement. Gracefully terminates workers when both the task and results queue is
-        empty, otherwise kills them without warning.
+        Enable the use of the ``with`` statement. Gracefully terminates workers when the task queue is not None.
         """
-        if self.tasks_queue is not None and (not self.tasks_queue.empty() or not self.results_queue.empty()):
+        if self._tasks_queue is not None:
             self.terminate()
-        else:
-            self.stop_and_join()
 
     def map(self, func_pointer, iterable_of_args, iterable_len=None, max_tasks_active=None, chunk_size=None,
-            restart_workers=False, worker_lifespan=None, progress_bar=False):
+            restart_workers=None, worker_lifespan=None, progress_bar=False):
         """
         Same as ``multiprocessing.map()``. Also allows a user to set the maximum number of tasks available in the queue.
         Note that this function can be slower than the unordered version.
@@ -388,9 +651,10 @@ class WorkerPool:
             the queue
         :param chunk_size: Int or ``None``. Number of simultaneous tasks to give to a worker. If ``None``, will generate
             ``n_jobs * 4`` number of chunks.
-        :param restart_workers: Boolean. Whether to restart the possibly already existing workers or use the old ones.
-            Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either way
-            when there are none.
+        :param restart_workers: DEPRECATED! MPIRE will no longer support reusing workers. This argument will be removed
+            from version 1.0.0 onwards. Boolean. Whether to restart the possibly already existing workers or use the old
+            ones. Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either
+            way when there are none.
         :param worker_lifespan: Int or ``None``. Number of chunks a worker can handle before it is restarted. If
             ``None``, workers will stay alive the entire time. Use this when workers use up too much memory over the
             course of time.
@@ -399,7 +663,7 @@ class WorkerPool:
         :return: List with ordered results
         """
         # Notify workers to keep order in mind
-        self.keep_order_event.set()
+        self._keep_order_event.set()
 
         # Process all args
         if iterable_len is None and hasattr(iterable_of_args, '__len__'):
@@ -409,13 +673,13 @@ class WorkerPool:
                                      progress_bar)
 
         # Notify workers to forget about order
-        self.keep_order_event.clear()
+        self._keep_order_event.clear()
 
         # Rearrange and return
         return [result[1] for result in sorted(results, key=lambda result: result[0])]
 
     def map_unordered(self, func_pointer, iterable_of_args, iterable_len=None, max_tasks_active=None, chunk_size=None,
-                      restart_workers=False, worker_lifespan=None, progress_bar=False):
+                      restart_workers=None, worker_lifespan=None, progress_bar=False):
         """
         Same as ``multiprocessing.map()``, but unordered. Also allows a user to set the maximum number of tasks
         available in the queue.
@@ -432,9 +696,10 @@ class WorkerPool:
             the queue
         :param chunk_size: Int or ``None``. Number of simultaneous tasks to give to a worker. If ``None``, will generate
             ``n_jobs * 4`` number of chunks.
-        :param restart_workers: Boolean. Whether to restart the possibly already existing workers or use the old ones.
-            Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either way
-            when there are none.
+        :param restart_workers: DEPRECATED! MPIRE will no longer support reusing workers. This argument will be removed
+            from version 1.0.0 onwards. Boolean. Whether to restart the possibly already existing workers or use the old
+            ones. Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either
+            way when there are none..
         :param worker_lifespan: Int or ``None``. Number of chunks a worker can handle before it is restarted. If
             ``None``, workers will stay alive the entire time. Use this when workers use up too much memory over the
             course of time.
@@ -447,7 +712,7 @@ class WorkerPool:
                                         restart_workers, worker_lifespan, progress_bar))
 
     def imap(self, func_pointer, iterable_of_args, iterable_len=None, max_tasks_active=None, chunk_size=None,
-             restart_workers=False, worker_lifespan=None, progress_bar=False):
+             restart_workers=None, worker_lifespan=None, progress_bar=False):
         """
         Same as ``multiprocessing.imap_unordered()``, but ordered. Also allows a user to set the maximum number of
         tasks available in the queue.
@@ -464,9 +729,10 @@ class WorkerPool:
             the queue
         :param chunk_size: Int or ``None``. Number of simultaneous tasks to give to a worker. If ``None``, will generate
             ``n_jobs * 4`` number of chunks.
-        :param restart_workers: Boolean. Whether to restart the possibly already existing workers or use the old ones.
-            Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either way
-            when there are none.
+        :param restart_workers: DEPRECATED! MPIRE will no longer support reusing workers. This argument will be removed
+            from version 1.0.0 onwards. Boolean. Whether to restart the possibly already existing workers or use the old
+            ones. Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either
+            way when there are none.
         :param worker_lifespan: Int or ``None``. Number of chunks a worker can handle before it is restarted. If
             ``None``, workers will stay alive the entire time. Use this when workers use up too much memory over the
             course of time.
@@ -475,7 +741,7 @@ class WorkerPool:
         :return: Generator yielding ordered results
         """
         # Notify workers to keep order in mind
-        self.keep_order_event.set()
+        self._keep_order_event.set()
 
         # Yield results in order
         next_result_idx = 0
@@ -507,10 +773,10 @@ class WorkerPool:
             yield tmp_results.pop(result_idx)
 
         # Notify workers to forget about order
-        self.keep_order_event.clear()
+        self._keep_order_event.clear()
 
     def imap_unordered(self, func_pointer, iterable_of_args, iterable_len=None, max_tasks_active=None, chunk_size=None,
-                       restart_workers=True, worker_lifespan=None, progress_bar=False):
+                       restart_workers=None, worker_lifespan=None, progress_bar=False):
         """
         Same as ``multiprocessing.imap_unordered()``. Also allows a user to set the maximum number of tasks available in
         the queue.
@@ -527,15 +793,108 @@ class WorkerPool:
             the queue
         :param chunk_size: Int or ``None``. Number of simultaneous tasks to give to a worker. If ``None``, will generate
             ``n_jobs * 4`` number of chunks.
-        :param restart_workers: Boolean. Whether to restart the possibly already existing workers or use the old ones.
-            Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either way
-            when there are none.
+        :param restart_workers: DEPRECATED! MPIRE will no longer support reusing workers. This argument will be removed
+            from version 1.0.0 onwards. Boolean. Whether to restart the possibly already existing workers or use the old
+            ones. Note: in the latter case the ``func_pointer`` parameter will have no effect. Will start workers either
+            way when there are none.
         :param worker_lifespan: Int or ``None``. Number of chunks a worker can handle before it is restarted. If
             ``None``, workers will stay alive the entire time. Use this when workers use up too much memory over the
             course of time.
         :param progress_bar: Boolean or ``tqdm`` instance. When ``True`` will display a default progress bar. Defaults
             can be overridden by supplying a custom ``tqdm`` progress bar instance. Use ``False`` to disable.
         :return: Generator yielding unordered results
+        """
+        # Check parameters and thereby obtain the number of tasks. The chunk_size and progress bar parameters could be
+        # modified as well
+        n_tasks, chunk_size, progress_bar = self._check_map_parameters(iterable_of_args, iterable_len, progress_bar,
+                                                                       chunk_size, max_tasks_active, worker_lifespan,
+                                                                       restart_workers)
+
+        # Reset some variables
+        self.exception_caught = False
+        self._cleaning_up = False
+        self._ready_for_destruction.clear()
+
+        # Chunk the function arguments
+        iterator_of_chunked_args = self.chunk_tasks(iterable_of_args, n_tasks, chunk_size)
+
+        # Create a progress bar if requested
+        progress_bar, progress_bar_progress = self._create_progress_bar(progress_bar, n_tasks)
+
+        # Start workers
+        self.start_workers(func_pointer, worker_lifespan)
+
+        # Start a thread that handles exceptions
+        self._exception_thread = Thread(target=self._handle_exceptions)
+        self._exception_thread.start()
+
+        # Process all args in the iterable. If maximum number of active tasks is None, we avoid all the if and
+        # try-except clauses to speed up the process.
+        n_active = 0
+        if max_tasks_active == 'n_jobs*2':
+            max_tasks_active = len(self._workers) * 2
+
+        if max_tasks_active is None:
+            for chunked_args in iterator_of_chunked_args:
+                # Stop given tasks when an exception was caught
+                if self.exception_caught:
+                    break
+
+                # Add task
+                self.add_task(chunked_args)
+                n_active += 1
+
+                # Restart workers if necessary
+                self._restart_workers()
+
+        elif isinstance(max_tasks_active, int):
+            while not self.exception_caught:
+                # Add task, only if allowed and if there are any
+                if n_active < max_tasks_active:
+                    try:
+                        self.add_task(next(iterator_of_chunked_args))
+                        n_active += 1
+                    except StopIteration:
+                        break
+
+                # Check if new results are available, but don't wait for it
+                try:
+                    yield from self._results_queue.get(block=False)
+                    self._results_queue.task_done()
+                    n_active -= 1
+                except queue.Empty:
+                    pass
+
+                # Restart workers if necessary
+                self._restart_workers()
+
+        # Obtain the results not yet obtained
+        while not self.exception_caught and n_active != 0:
+            try:
+                yield from self._results_queue.get(block=True, timeout=0.1)
+                self._results_queue.task_done()
+                n_active -= 1
+            except queue.Empty:
+                pass
+
+            # Restart workers if necessary
+            self._restart_workers()
+
+        # Clean up time. When we're done with the progress bar we're essentially ready for destruction in the case of an
+        # exception that was raised in a child process.
+        self._clean_up_progress_bar(progress_bar, progress_bar_progress)
+        self._ready_for_destruction.set()
+        self._clean_up_exception_handle()
+        self.stop_and_join()
+
+    @staticmethod
+    def _check_map_parameters(iterable_of_args, iterable_len, progress_bar, chunk_size, max_tasks_active,
+                              worker_lifespan, restart_workers):
+        """
+        Check the parameters provided to any (i)map function. Also extracts the number of tasks and can modify the
+        ``chunk_size`` and ``progress_bar`` parameters.
+
+        :return: Number of tasks, chunk size, progress bar
         """
         # Get number of tasks
         n_tasks = None
@@ -554,91 +913,31 @@ class WorkerPool:
             chunk_size = 1
             progress_bar = False
 
+        # Check max_tasks_active parameter
+        if isinstance(max_tasks_active, int):
+            if max_tasks_active <= 0:
+                raise ValueError('mmax_tasks_active should be a positive integer, None or "n_jobs*2')
+        elif max_tasks_active is not None and max_tasks_active is not 'n_jobs*2':
+            raise TypeError('max_tasks_active should be a positive integer, None or "n_jobs*2"')
+
+        # If worker lifespan is not None or not a positive integer, raise
+        if isinstance(worker_lifespan, int):
+            if worker_lifespan <= 0:
+                raise ValueError('worker_lifespan should be either None or a positive integer (> 0)')
+        elif worker_lifespan is not None:
+            raise TypeError('worker_lifespan should be either None or a positive integer (> 0)')
+
         # Parameter total should be given
         if isinstance(progress_bar, tqdm.tqdm) and progress_bar.total is None:
             raise ValueError("Custom tqdm progress bar instance needs parameter 'total'")
 
-        # Chunk the function arguments
-        iterator_of_chunked_args = self.chunk_tasks(iterable_of_args, n_tasks, chunk_size)
+        # Deprecated feature
+        if restart_workers is not None:
+            warnings.simplefilter('default')
+            warnings.warn('Argument restart_workers is deprecated. MPIRE will no longer support reusing workers. This '
+                          'argument will be removed from version 1.0.0 onwards.', DeprecationWarning, stacklevel=2)
 
-        # Create a progress bar if requested
-        if progress_bar is True or isinstance(progress_bar, tqdm.tqdm):
-            progress_bar = progress_bar if isinstance(progress_bar, tqdm.tqdm) else tqdm_pb(total=n_tasks)
-            self.task_completed_queue = Queue()
-
-            # Start a thread
-            progress_bar_thread = Thread(target=self._handle_progress_bar, args=(progress_bar, n_tasks))
-            progress_bar_thread.start()
-        else:
-            self.task_completed_queue = None
-            progress_bar_thread = None
-
-        # Start workers
-        if not self.workers or restart_workers:
-            self.start_workers(func_pointer, worker_lifespan)
-
-        # Process all args in the iterable. If maximum number of active tasks is None, we avoid all the if and
-        # try-except clauses to speed up the process.
-        n_active = 0
-        if max_tasks_active == 'n_jobs*2':
-            max_tasks_active = len(self.workers) * 2
-        if max_tasks_active is None:
-            for chunked_args in iterator_of_chunked_args:
-                self.add_task(chunked_args)
-                n_active += 1
-
-                # Restart workers if necessary
-                self._restart_workers()
-        elif isinstance(max_tasks_active, int) and max_tasks_active > 0:
-            while True:
-                # Add task, only if allowed and if there are any
-                if n_active < max_tasks_active:
-                    try:
-                        self.add_task(next(iterator_of_chunked_args))
-                        n_active += 1
-                    except StopIteration:
-                        break
-
-                # Check if new results are available, but don't wait for it
-                try:
-                    yield from self.results_queue.get(block=False)
-                    n_active -= 1
-                except queue.Empty:
-                    pass
-
-                # Restart workers if necessary
-                self._restart_workers()
-        else:
-            raise ValueError('Maximum number of active tasks must be at least 1')
-
-        # Obtain the results not yet obtained
-        while n_active != 0:
-            try:
-                yield from self.results_queue.get(block=True, timeout=0.1)
-                n_active -= 1
-            except queue.Empty:
-                pass
-
-            # Restart workers if necessary
-            self._restart_workers()
-
-        # Clean up the progress bar
-        if progress_bar:
-            progress_bar_thread.join()
-            progress_bar.close()
-
-    def _handle_progress_bar(self, progress_bar, n_tasks):
-        """
-        When a task is completed a single item is put in the designated queue. This function, living in a thread, waits
-        for items in the queue and updates the progress bar accordingly.
-
-        :param progress_bar: tqdm progress bar object.
-        :param n_tasks: Int. Number of tasks to be expected.
-        """
-        for _ in range(n_tasks):
-            # Wait for job to finish
-            self.task_completed_queue.get(block=True)
-            progress_bar.update(1)
+        return n_tasks, chunk_size, progress_bar
 
     def chunk_tasks(self, iterable_of_args, n_tasks, chunk_size=None):
         """
@@ -654,7 +953,7 @@ class WorkerPool:
         """
         # Determine chunk size
         if chunk_size is None:
-            chunk_size, extra = divmod(n_tasks, len(self.workers) * 4)
+            chunk_size, extra = divmod(n_tasks, len(self._workers) * 4)
             if extra:
                 chunk_size += 1
 
@@ -665,3 +964,130 @@ class WorkerPool:
             if not chunk:
                 return
             yield chunk
+
+    def _create_progress_bar(self, progress_bar, n_tasks):
+        """
+        When provided, creates a progress bar and starts a thread that updates it. Otherwise, does nothing
+
+        :param progress_bar: Boolean or ``tqdm`` instance. When ``True`` will display a default progress bar. Defaults
+            can be overridden by supplying a custom ``tqdm`` progress bar instance. Use ``False`` to disable.
+        :param n_tasks: Int. Number of tasks we have to process.
+        :return: False or ``tqdm`` instance.
+        """
+        # Create a progress bar if requested
+        if progress_bar is True or isinstance(progress_bar, tqdm.tqdm):
+            progress_bar = progress_bar if isinstance(progress_bar, tqdm.tqdm) else tqdm_pb(total=n_tasks)
+            progress_bar_progress = Value('i', 0, lock=True)
+            self._task_completed_queue = JoinableQueue()
+
+            # Start a thread
+            self._progress_bar_thread = Thread(target=self._handle_progress_bar,
+                                               args=(progress_bar, progress_bar_progress))
+            self._progress_bar_thread.start()
+        else:
+            self._task_completed_queue = None
+            self._progress_bar_thread = None
+            progress_bar_progress = None
+
+        return progress_bar, progress_bar_progress
+
+    def _handle_progress_bar(self, progress_bar, progress_bar_progress):
+        """
+        When a task is completed a single item is put in the designated queue. This function, living in a thread, waits
+        for items in the queue and updates the progress bar accordingly.
+
+        :param progress_bar: tqdm progress bar object.
+        :param progress_bar_progress: multiprocessing.Value for keeping track of the number of updates
+        """
+        while True:
+            # Wait for a job to finish
+            task_completed = self._task_completed_queue.get(block=True)
+
+            # If we received None, we should quit right away
+            if task_completed is None:
+                self._task_completed_queue.task_done()
+                break
+
+            # Update progress bar. Note that we also update a separate counter which is used to check if the progress
+            # bar is completed. I realize that tqdm has a public variable `n` which should keep track of the current
+            # progress, but for some reason that variable doesn't work here, it equals the `total` variable all the
+            # time.
+            progress_bar.update(1)
+            self._task_completed_queue.task_done()
+            progress_bar_progress.value += 1
+
+    def _clean_up_progress_bar(self, progress_bar, progress_bar_progress):
+        """
+        Closes the progress bar and its accompanying thread if a progress bar was provided.
+
+        :param progress_bar: False or ``tqdm`` instance
+        """
+        # Clean up the progress bar
+        if progress_bar:
+            # Wait until all tasks are completed (well, not when an exception was caught). We have to wait because some
+            # queues can get delayed. It can occur that the results have already been processed and that this function
+            # is called, before all the task_completed items have been received. This means that we would put a poison
+            # pill in the task_completed_queue while there are still items to process. This results in a deadlock when
+            # we want to join the task_completed_queue on a later point.
+            while not self.exception_caught and progress_bar_progress.value != progress_bar.total:
+                pass
+
+            # Insert poison pill and close the progress bar and its handling thread
+            self._task_completed_queue.put(None)
+            self._progress_bar_thread.join()
+            self._progress_bar_thread = None
+            progress_bar.close()
+
+    def _handle_exceptions(self):
+        """
+        Checks for raised exceptions by child processes.
+        """
+        # Wait for an exception to occur
+        err, traceback_str = self._exception_queue.get(block=True)
+
+        # If we received None, we should just quit quietly
+        if err is not None:
+            # Let main process know we can stop working
+            self.exception_caught = True
+
+            # Wait until the main process realizes it should stop. We have to wait because the main process could be
+            # right in the middle of adding tasks to queues. If we terminate and drain a queue while at the same time it
+            # is being written to we can get errors.
+            self._ready_for_destruction.wait()
+
+            # Kill processes
+            self.terminate(keep_exception_queue=True)
+
+            # Make sure we only process a single exception (multiple processes could potentially insert exceptions at
+            # the same time). Note that draining the queue makes sure that task_done() is called for the above 'get'
+            # call
+            self._drain_and_join_queue(self._exception_queue, join=False)
+
+            # Pass error to main process so it can be raised there (exceptions raised from threads or child processes
+            # cannot be caught directly)
+            self._exception_queue.put((err, traceback_str))
+        else:
+            self._exception_queue.task_done()
+
+    def _clean_up_exception_handle(self):
+        """
+        Closes the exception handling thread and re-raises any exception encountered there.
+        """
+        # Shutdown the exception handling thread. Insert a poison pill when no exception was raised by the workers.
+        # If there was an exception, the exception handling thread would already be joinable.
+        if not self.exception_caught:
+            self._exception_queue.put((None, None))
+        self._exception_thread.join()
+        self._exception_thread = None
+
+        # If there was any exception the exception handling thread passed on the error to the main process, so we can
+        # raise it here.
+        if self.exception_caught:
+            err, traceback_str = self._exception_queue.get(block=True)
+            self._exception_queue.task_done()
+            self._exception_queue.join()
+            self._exception_queue = None
+            raise err(traceback_str)
+        else:
+            self._exception_queue.join()
+            self._exception_queue = None